fix compiler warnings

+++ Windows/Gopher/Gopher.h
@@ -18,8 +18,8 @@
 class Gopher
 {
 private:
-  int DEAD_ZONE = 6000;                 // Thumbstick dead zone to use for mouse movement. Absolute maximum shall be 65534.
-  int SCROLL_DEAD_ZONE = 5000;          // Thumbstick dead zone to use for scroll wheel movement. Absolute maximum shall be 65534.
+  float DEAD_ZONE = 6000;                 // Thumbstick dead zone to use for mouse movement. Absolute maximum shall be 65534.
+  float SCROLL_DEAD_ZONE = 5000;          // Thumbstick dead zone to use for scroll wheel movement. Absolute maximum shall be 65534.
   int TRIGGER_DEAD_ZONE = 0;            // Dead zone for the left and right triggers to detect a trigger press. 0 means that any press to trigger will be read as a button press.
   float SCROLL_SPEED = 0.1f;             // Speed at which you scroll.
   const int FPS = 150;                  // Update rate of the main Gopher loop. Interpreted as cycles-per-second.
@@ -40,7 +40,7 @@ private:
   float _yRest = 0.0f;
 
   bool _disabled = false;           // Disables the Gopher controller mapping.
-  bool _vibrationDisabled = false;  // Prevents Gopher from producing controller vibrations. 
+  bool _vibrationDisabled = false;  // Prevents Gopher from producing controller vibrations.
   bool _hidden = false;             // Gopher main window visibility.
   bool _lTriggerPrevious = false;   // Previous state of the left trigger.
   bool _rTriggerPrevious = false;   // Previous state of the right trigger.
@@ -53,7 +53,7 @@ private:
   DWORD CONFIG_MOUSE_LEFT = NULL;
   DWORD CONFIG_MOUSE_RIGHT = NULL;
   DWORD CONFIG_MOUSE_MIDDLE = NULL;
-  
+
   // Gopher Settings
   DWORD CONFIG_HIDE = NULL;
   DWORD CONFIG_DISABLE = NULL;
@@ -62,22 +62,22 @@ private:
   DWORD CONFIG_OSK = NULL;
 
   // Gamepad bindings
-  DWORD GAMEPAD_DPAD_UP = NULL;
-  DWORD GAMEPAD_DPAD_DOWN = NULL;
-  DWORD GAMEPAD_DPAD_LEFT = NULL;
-  DWORD GAMEPAD_DPAD_RIGHT = NULL;
-  DWORD GAMEPAD_START = NULL;
-  DWORD GAMEPAD_BACK = NULL;
-  DWORD GAMEPAD_LEFT_THUMB = NULL;
-  DWORD GAMEPAD_RIGHT_THUMB = NULL;
-  DWORD GAMEPAD_LEFT_SHOULDER = NULL;
-  DWORD GAMEPAD_RIGHT_SHOULDER = NULL;
-  DWORD GAMEPAD_A = NULL;
-  DWORD GAMEPAD_B = NULL;
-  DWORD GAMEPAD_X = NULL;
-  DWORD GAMEPAD_Y = NULL;
-  DWORD GAMEPAD_TRIGGER_LEFT = NULL;
-  DWORD GAMEPAD_TRIGGER_RIGHT = NULL;
+  WORD GAMEPAD_DPAD_UP = NULL;
+  WORD GAMEPAD_DPAD_DOWN = NULL;
+  WORD GAMEPAD_DPAD_LEFT = NULL;
+  WORD GAMEPAD_DPAD_RIGHT = NULL;
+  WORD GAMEPAD_START = NULL;
+  WORD GAMEPAD_BACK = NULL;
+  WORD GAMEPAD_LEFT_THUMB = NULL;
+  WORD GAMEPAD_RIGHT_THUMB = NULL;
+  WORD GAMEPAD_LEFT_SHOULDER = NULL;
+  WORD GAMEPAD_RIGHT_SHOULDER = NULL;
+  WORD GAMEPAD_A = NULL;
+  WORD GAMEPAD_B = NULL;
+  WORD GAMEPAD_X = NULL;
+  WORD GAMEPAD_Y = NULL;
+  WORD GAMEPAD_TRIGGER_LEFT = NULL;
+  WORD GAMEPAD_TRIGGER_RIGHT = NULL;
 
   // Button press state logic variables
   std::map<DWORD, bool> _xboxClickStateLastIteration;
